refactor internal interface, improve toplevel expression hoisting:
- create a simple internal interface and allows each profiling point to 
define its reports
- remove unnecessities, simplify logics
